Move all code into ratingrelay.py, remove other *.py files
